v4l2src: Support num-alloc-buffer option
This option uses to decide number of buffers will allocate.
And it only supports for io-mode=mmap/dmabuf
Default is auto decision, it will be calculated based on requirement
from driver, proposal from downstream element and so on.
If the value of num-alloc-buffer is too small or too big, number
of buffers will be adjusted to suit with supported range of driver.

refs #149515
Signed-off-by: Thao Luong <thao.luong.xh@renesas.com>

porting from 1.12.2 to 1.14.2
Signed-off-by: Shinya Saito <ssaito@igel.co.jp>