v4l2bufferpool: mmngr: Create video meta data for mmngr buffers
The video meta data for the mmngr buffers that wrap V4L2 buffers
cannot be copied from the V4L2 buffers.  In general, video meta data
can only be copied if the buffer memory is copied with it. Instead,
a new meta data structure must be created with the necessary informantion.

Timestamp and flag data can and should be copied, so make sure that is
done too.